requirements.txt
pystray>=0.19.4
Pillow>=9.5.0
